+++ lib/simple_form/inputs/collection_select_input.rb
@@ -1,0 +1,14 @@
+module SimpleForm
+  module Inputs
+    class CollectionSelectInput < CollectionInput
+      def input
+        label_method, value_method = detect_collection_methods
+
+        @builder.collection_select(
+          attribute_name, collection, value_method, label_method,
+          input_options, input_html_options
+        )
+      end
+    end
+  end
+end