+++ test/inputs/collection_select_input_test.rb
@@ -1,36 +1,9 @@
 # encoding: UTF-8
 require 'test_helper'
 
-class CollectionInputTest < ActionView::TestCase
+class CollectionSelectInputTest < ActionView::TestCase
   setup do
-    SimpleForm::Inputs::CollectionInput.reset_i18n_cache :boolean_collection
-  end
-
-  test 'input should generate boolean radio buttons by default for radio types' do
-    with_input_for @user, :active, :radio
-    assert_select 'input[type=radio][value=true].radio#user_active_true'
-    assert_select 'input[type=radio][value=false].radio#user_active_false'
-  end
-
-  test 'input as radio should generate internal labels by default' do
-    with_input_for @user, :active, :radio
-    assert_select 'label[for=user_active_true]', 'Yes'
-    assert_select 'label[for=user_active_false]', 'No'
-  end
-
-  test 'input as radio should use i18n to translate internal labels' do
-    store_translations(:en, :simple_form => { :yes => 'Sim', :no => 'Não' }) do
-      with_input_for @user, :active, :radio
-      assert_select 'label[for=user_active_true]', 'Sim'
-      assert_select 'label[for=user_active_false]', 'Não'
-    end
-  end
-
-  test 'input should mark the checked value when using boolean and radios' do
-    @user.active = false
-    with_input_for @user, :active, :radio
-    assert_no_select 'input[type=radio][value=true][checked]'
-    assert_select 'input[type=radio][value=false][checked]'
+    SimpleForm::Inputs::CollectionSelectInput.reset_i18n_cache :boolean_collection
   end
 
   test 'input should generate a boolean select with options by default for select types' do
@@ -161,96 +134,6 @@ class CollectionInputTest < ActionView::TestCase
     assert_no_select 'div.disabled'
   end
 
-  test 'input should allow overriding collection for radio types' do
-    with_input_for @user, :name, :radio, :collection => ['Jose', 'Carlos']
-    assert_select 'input[type=radio][value=Jose]'
-    assert_select 'input[type=radio][value=Carlos]'
-    assert_select 'label.collection_radio', 'Jose'
-    assert_select 'label.collection_radio', 'Carlos'
-  end
-
-  test 'input should do automatic collection translation for radio types using defaults key' do
-    store_translations(:en, :simple_form => { :options => { :defaults => {
-      :gender => { :male => 'Male', :female => 'Female'}
-    } } } ) do
-      with_input_for @user, :gender, :radio, :collection => [:male, :female]
-      assert_select 'input[type=radio][value=male]'
-      assert_select 'input[type=radio][value=female]'
-      assert_select 'label.collection_radio', 'Male'
-      assert_select 'label.collection_radio', 'Female'
-    end
-  end
-
-  test 'input should do automatic collection translation for radio types using specific object key' do
-    store_translations(:en, :simple_form => { :options => { :user => {
-      :gender => { :male => 'Male', :female => 'Female'}
-    } } } ) do
-      with_input_for @user, :gender, :radio, :collection => [:male, :female]
-      assert_select 'input[type=radio][value=male]'
-      assert_select 'input[type=radio][value=female]'
-      assert_select 'label.collection_radio', 'Male'
-      assert_select 'label.collection_radio', 'Female'
-    end
-  end
-
-  test 'input should mark the current radio value by default' do
-    @user.name = "Carlos"
-    with_input_for @user, :name, :radio, :collection => ['Jose', 'Carlos']
-    assert_select 'input[type=radio][value=Carlos][checked=checked]'
-  end
-
-  test 'input should allow using a collection with text/value arrays' do
-    with_input_for @user, :name, :radio, :collection => [['Jose', 'jose'], ['Carlos', 'carlos']]
-    assert_select 'input[type=radio][value=jose]'
-    assert_select 'input[type=radio][value=carlos]'
-    assert_select 'label.collection_radio', 'Jose'
-    assert_select 'label.collection_radio', 'Carlos'
-  end
-
-  test 'input should allow using a collection with a Proc' do
-    with_input_for @user, :name, :radio, :collection => Proc.new { ['Jose', 'Carlos' ] }
-    assert_select 'label.collection_radio', 'Jose'
-    assert_select 'label.collection_radio', 'Carlos'
-  end
-
-  test 'input should allow overriding only label method for collections' do
-    with_input_for @user, :name, :radio,
-                          :collection => ['Jose' , 'Carlos'],
-                          :label_method => :upcase
-    assert_select 'label.collection_radio', 'JOSE'
-    assert_select 'label.collection_radio', 'CARLOS'
-  end
-
-  test 'input should allow overriding only value method for collections' do
-    with_input_for @user, :name, :radio,
-                          :collection => ['Jose' , 'Carlos'],
-                          :value_method => :upcase
-    assert_select 'input[type=radio][value=JOSE]'
-    assert_select 'input[type=radio][value=CARLOS]'
-  end
-
-  test 'input should allow overriding label and value method for collections' do
-    with_input_for @user, :name, :radio,
-                          :collection => ['Jose' , 'Carlos'],
-                          :label_method => :upcase,
-                          :value_method => :downcase
-    assert_select 'input[type=radio][value=jose]'
-    assert_select 'input[type=radio][value=carlos]'
-    assert_select 'label.collection_radio', 'JOSE'
-    assert_select 'label.collection_radio', 'CARLOS'
-  end
-
-  test 'input should allow overriding label and value method using a lambda for collections' do
-    with_input_for @user, :name, :radio,
-                          :collection => ['Jose' , 'Carlos'],
-                          :label_method => lambda { |i| i.upcase },
-                          :value_method => lambda { |i| i.downcase }
-    assert_select 'input[type=radio][value=jose]'
-    assert_select 'input[type=radio][value=carlos]'
-    assert_select 'label.collection_radio', 'JOSE'
-    assert_select 'label.collection_radio', 'CARLOS'
-  end
-
   test 'input should allow overriding label and value method using a lambda for collection selects' do
     with_input_for @user, :name, :select,
                           :collection => ['Jose' , 'Carlos'],
@@ -283,24 +166,10 @@ class CollectionInputTest < ActionView::TestCase
     assert_select 'select option[value=carlos]', 'carlos'
   end
 
-  test 'collection input with radio type should generate required html attribute' do
-    with_input_for @user, :name, :radio, :collection => ['Jose' , 'Carlos']
-    assert_select 'input[type=radio].required'
-    assert_select 'input[type=radio][required]'
-  end
-
   test 'collection input with select type should generate required html attribute only with blank option' do
     with_input_for @user, :name, :select, :include_blank => true, :collection => ['Jose' , 'Carlos']
     assert_select 'select.required'
     assert_select 'select[required]'
-  end
-
-  test 'when not using browser validations, input should not generate required html attribute' do
-    swap SimpleForm, :browser_validations => false do
-      with_input_for @user, :name, :string
-      assert_select 'input[type=text].required'
-      assert_no_select 'input[type=text][required]'
-    end
   end
 
   test 'collection input with select type should not generate required html attribute without blank option' do
@@ -319,36 +188,6 @@ class CollectionInputTest < ActionView::TestCase
     with_input_for @user, :name, :select, :include_blank => true, :input_html => {:multiple => true}, :collection => ['Jose' , 'Carlos']
     assert_select 'select.required'
     assert_select 'select[required]'
-  end
-
-  test 'collection input with check_boxes type should not generate required html attribute' do
-    with_input_for @user, :name, :check_boxes, :collection => ['Jose' , 'Carlos']
-    assert_select 'input.required'
-    assert_no_select 'input[required]'
-  end
-
-  test 'input should do automatic collection translation for check_box types using defaults key' do
-    store_translations(:en, :simple_form => { :options => { :defaults => {
-      :gender => { :male => 'Male', :female => 'Female'}
-    } } } ) do
-      with_input_for @user, :gender, :check_boxes, :collection => [:male, :female]
-      assert_select 'input[type=checkbox][value=male]'
-      assert_select 'input[type=checkbox][value=female]'
-      assert_select 'label.collection_check_boxes', 'Male'
-      assert_select 'label.collection_check_boxes', 'Female'
-    end
-  end
-
-  test 'input should do automatic collection translation for check_box types using specific object key' do
-    store_translations(:en, :simple_form => { :options => { :user => {
-      :gender => { :male => 'Male', :female => 'Female'}
-    } } } ) do
-      with_input_for @user, :gender, :check_boxes, :collection => [:male, :female]
-      assert_select 'input[type=checkbox][value=male]'
-      assert_select 'input[type=checkbox][value=female]'
-      assert_select 'label.collection_check_boxes', 'Male'
-      assert_select 'label.collection_check_boxes', 'Female'
-    end
   end
 
   test 'input should allow disabled options with a lambda for collection select' do